chore: Secretrefs refactoring, first iteration
