+++ .golangci.yaml
@@ -116,9 +116,11 @@ linters-settings:
         alias: metricsserver
       - pkg: sigs.k8s.io/controller-runtime/pkg/predicate
         alias: ctrlpredicate
-      - pkg: "github.com/kyma-project/telemetry-manager/internal/reconciler/commonstatus/stubs"
+      - pkg: github.com/kyma-project/telemetry-manager/utils/(\w+)
+        alias: utils
+      - pkg: github.com/kyma-project/telemetry-manager/internal/reconciler/commonstatus/stubs
         alias: commonStatusStubs
-      - pkg: "k8s.io/client-go/testing"
+      - pkg: k8s.io/client-go/testing
         alias: clienttesting
   errcheck:
     check-type-assertions: true # Reports type assertions: `a := b.(SomeStruct)`.
