+++ internal/reconciler/logpipeline/fluentbit/sync.go
@@ -14,6 +14,7 @@ import (
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/telemetry-manager/internal/fluentbit/config/builder"
 	"github.com/kyma-project/telemetry-manager/internal/k8sutils"
+	utils "github.com/kyma-project/telemetry-manager/utils/logpipeline"
 )
 
 type syncer struct {
@@ -141,7 +142,7 @@ func (s *syncer) syncEnvSecret(ctx context.Context, logPipelines []telemetryv1al
 			continue
 		}
 
-		for _, ref := range logPipelines[i].GetEnvSecretRefs() {
+		for _, ref := range utils.GetEnvSecretRefs(&logPipelines[i]) {
 			targetKey := builder.FormatEnvVarName(logPipelines[i].Name, ref.Namespace, ref.Name, ref.Key)
 			if copyErr := s.copySecretData(ctx, ref, targetKey, newSecret.Data); copyErr != nil {
 				return fmt.Errorf("unable to copy secret data: %w", copyErr)
