+++ internal/reconciler/logpipeline/fluentbit/validator.go
@@ -5,8 +5,8 @@ import (
 
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/telemetry-manager/internal/validators/endpoint"
-	"github.com/kyma-project/telemetry-manager/internal/validators/secretref"
 	"github.com/kyma-project/telemetry-manager/internal/validators/tlscert"
+	utils "github.com/kyma-project/telemetry-manager/utils/logpipeline"
 )
 
 type EndpointValidator interface {
@@ -18,7 +18,7 @@ type TLSCertValidator interface {
 }
 
 type SecretRefValidator interface {
-	Validate(ctx context.Context, getter secretref.Getter) error
+	Validate(ctx context.Context, refs []telemetryv1alpha1.SecretKeyRef) error
 }
 
 type Validator struct {
@@ -28,7 +28,7 @@ type Validator struct {
 }
 
 func (v *Validator) validate(ctx context.Context, pipeline *telemetryv1alpha1.LogPipeline) error {
-	if err := v.SecretRefValidator.Validate(ctx, pipeline); err != nil {
+	if err := v.SecretRefValidator.Validate(ctx, utils.GetSecretRefs(pipeline)); err != nil {
 		return err
 	}
