+++ internal/reconciler/tracepipeline/validator.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
-	"github.com/kyma-project/telemetry-manager/internal/validators/secretref"
 	"github.com/kyma-project/telemetry-manager/internal/validators/tlscert"
+	utils "github.com/kyma-project/telemetry-manager/utils/tracepipeline"
 )
 
 type EndpointValidator interface {
@@ -13,7 +13,7 @@ type EndpointValidator interface {
 }
 
 type SecretRefValidator interface {
-	Validate(ctx context.Context, getter secretref.Getter) error
+	Validate(ctx context.Context, refs []telemetryv1alpha1.SecretKeyRef) error
 }
 
 type TLSCertValidator interface {
@@ -28,7 +28,7 @@ type Validator struct {
 }
 
 func (v *Validator) validate(ctx context.Context, pipeline *telemetryv1alpha1.TracePipeline) error {
-	if err := v.SecretRefValidator.Validate(ctx, pipeline); err != nil {
+	if err := v.SecretRefValidator.Validate(ctx, utils.GetSecretRefs(pipeline)); err != nil {
 		return err
 	}
