+++ internal/validators/secretref/secret_ref_test.go
@@ -12,23 +12,15 @@ import (
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 )
 
-type mockGetter struct {
-	refs []telemetryv1alpha1.SecretKeyRef
-}
-
-func (m mockGetter) GetSecretRefs() []telemetryv1alpha1.SecretKeyRef {
-	return m.refs
-}
-
 func TestValidate(t *testing.T) {
 	tests := []struct {
 		name        string
-		getter      []telemetryv1alpha1.SecretKeyRef
+		refs        []telemetryv1alpha1.SecretKeyRef
 		expectError error
 	}{
 		{
 			name: "Success",
-			getter: []telemetryv1alpha1.SecretKeyRef{
+			refs: []telemetryv1alpha1.SecretKeyRef{
 				{Name: "my-secret1", Namespace: "default", Key: "myKey1"},
 				{Name: "my-secret2", Namespace: "default", Key: "myKey2"},
 			},
@@ -36,7 +28,7 @@ func TestValidate(t *testing.T) {
 		},
 		{
 			name: "SecretNameNotPresent",
-			getter: []telemetryv1alpha1.SecretKeyRef{
+			refs: []telemetryv1alpha1.SecretKeyRef{
 				{Name: "my-secret1", Namespace: "default", Key: "myKey1"},
 				{Name: "notExistent", Namespace: "default", Key: "myKey2"},
 			},
@@ -44,7 +36,7 @@ func TestValidate(t *testing.T) {
 		},
 		{
 			name: "SecretNamespaceNotPresent",
-			getter: []telemetryv1alpha1.SecretKeyRef{
+			refs: []telemetryv1alpha1.SecretKeyRef{
 				{Name: "my-secret1", Namespace: "default", Key: "myKey1"},
 				{Name: "my-secret2", Namespace: "notExistent", Key: "myKey2"},
 			},
@@ -52,7 +44,7 @@ func TestValidate(t *testing.T) {
 		},
 		{
 			name: "SecretKeyNotPresent",
-			getter: []telemetryv1alpha1.SecretKeyRef{
+			refs: []telemetryv1alpha1.SecretKeyRef{
 				{Name: "my-secret1", Namespace: "default", Key: "myKey1"},
 				{Name: "my-secret2", Namespace: "default", Key: "notExistent"},
 			},
@@ -83,14 +75,10 @@ func TestValidate(t *testing.T) {
 
 			client := fake.NewClientBuilder().WithObjects(&existingSecret1).WithObjects(&existingSecret2).Build()
 
-			getter := mockGetter{
-				refs: test.getter,
-			}
-
 			secretRefValidator := Validator{
 				Client: client,
 			}
-			err := secretRefValidator.Validate(context.TODO(), getter)
+			err := secretRefValidator.Validate(context.TODO(), test.refs)
 			require.ErrorIs(t, err, test.expectError)
 		})
 	}
@@ -99,13 +87,13 @@ func TestValidate(t *testing.T) {
 func TestGetValue(t *testing.T) {
 	tests := []struct {
 		name          string
-		getter        telemetryv1alpha1.SecretKeyRef
+		refs          telemetryv1alpha1.SecretKeyRef
 		expectError   error
 		expectedValue string
 	}{
 		{
 			name: "Success",
-			getter: telemetryv1alpha1.SecretKeyRef{
+			refs: telemetryv1alpha1.SecretKeyRef{
 				Name: "my-secret1", Namespace: "default", Key: "myKey1",
 			},
 			expectError:   nil,
@@ -113,7 +101,7 @@ func TestGetValue(t *testing.T) {
 		},
 		{
 			name: "SecretNameNotPresent",
-			getter: telemetryv1alpha1.SecretKeyRef{
+			refs: telemetryv1alpha1.SecretKeyRef{
 				Name: "notExistent", Namespace: "default", Key: "myKey1",
 			},
 			expectError:   ErrSecretRefNotFound,
@@ -121,7 +109,7 @@ func TestGetValue(t *testing.T) {
 		},
 		{
 			name: "SecretNamespaceNotPresent",
-			getter: telemetryv1alpha1.SecretKeyRef{
+			refs: telemetryv1alpha1.SecretKeyRef{
 				Name: "my-secret1", Namespace: "notExistent", Key: "myKey1",
 			},
 			expectError:   ErrSecretRefNotFound,
@@ -129,7 +117,7 @@ func TestGetValue(t *testing.T) {
 		},
 		{
 			name: "SecretKeyNotPresent",
-			getter: telemetryv1alpha1.SecretKeyRef{
+			refs: telemetryv1alpha1.SecretKeyRef{
 				Name: "my-secret1", Namespace: "default", Key: "notExistent",
 			},
 			expectError:   ErrSecretKeyNotFound,
@@ -137,7 +125,7 @@ func TestGetValue(t *testing.T) {
 		},
 		{
 			name: "SecretRefMissingKey",
-			getter: telemetryv1alpha1.SecretKeyRef{
+			refs: telemetryv1alpha1.SecretKeyRef{
 				Name: "my-secret1", Namespace: "default",
 			},
 			expectError:   ErrSecretRefMissingFields,
@@ -145,7 +133,7 @@ func TestGetValue(t *testing.T) {
 		},
 		{
 			name: "SecretRefMissingName",
-			getter: telemetryv1alpha1.SecretKeyRef{
+			refs: telemetryv1alpha1.SecretKeyRef{
 				Namespace: "default", Key: "notExistent",
 			},
 			expectError:   ErrSecretRefMissingFields,
@@ -153,7 +141,7 @@ func TestGetValue(t *testing.T) {
 		},
 		{
 			name: "SecretRefMissingNamespace",
-			getter: telemetryv1alpha1.SecretKeyRef{
+			refs: telemetryv1alpha1.SecretKeyRef{
 				Name: "my-secret1", Key: "notExistent",
 			},
 			expectError:   ErrSecretRefMissingFields,
@@ -175,7 +163,7 @@ func TestGetValue(t *testing.T) {
 
 			client := fake.NewClientBuilder().WithObjects(&existingSecret1).Build()
 
-			result, err := GetValue(context.TODO(), client, test.getter)
+			result, err := GetValue(context.TODO(), client, test.refs)
 
 			require.Equal(t, test.expectedValue, string(result))
 			require.ErrorIs(t, err, test.expectError)
